Add category andAdd category and general changes general changes
Complete the product and category section [make the fields dynamic]
Separate files and create utils
Improving the serialization of products
Add product visit section by checking ip

+++ Store/products/signals.py
@@ -1,12 +1,22 @@
 from django.dispatch import receiver
-from django.db.models.signals import post_delete
+from django.db.models.signals import post_delete,pre_save
 import os
+from .models import Ip
 
 
 @receiver(post_delete,sender="products.Product")
 def run_after_delete(instance,**kwargs):
+    
+    if instance.image:
+        path = instance.image.path
 
-    path = instance.image.path
+        if os.path.exists(path):
+            os.remove(path)
 
-    if os.path.exists(path):
-        os.remove(path)
+    instance.ip.delete()
+
+
+@receiver(pre_save,sender="products.Product")
+def run_after_Save(instance,**kwargs):
+    if not instance.ip:
+        instance.ip = Ip.objects.create(id=instance.id,title=instance.title)
